+++ src/components/ReactTypeformEmbed/components/Popup/Popup.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import { createPopup } from '@typeform/embed';
+import Modal from '../Modal';
+import '@typeform/embed/build/css/popup.css';
+
+const Popup = (props) => <Modal create={createPopup} {...props} />;
+
+export default Popup;