+++ src/components/ReactTypeformEmbed/components/Sidetab/index.js
@@ -1,0 +1,3 @@
+import Sidetab from './Sidetab';
+
+export default Sidetab;