+++ src/components/ReactTypeformEmbed/components/index.js
@@ -1,0 +1,7 @@
+import Widget from './Widget';
+import Popup from './Popup';
+import Slider from './Slider';
+import Sidetab from './Sidetab';
+import Popover from './Popover';
+
+export { Widget, Popup, Slider, Sidetab, Popover };