+++ src/components/index.js
@@ -1,1 +1,3 @@
-export { default as ReactTypeformEmbed } from './ReactTypeformEmbed';
+import { Widget, Popup, Slider, Sidetab, Popover } from './ReactTypeformEmbed';
+
+export { Widget, Popup, Slider, Sidetab, Popover };