+++ src/examples/components/ExamplePopover/ExamplePopover.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import { Popover } from '../../../components';
+
+const ExamplePopover = () => (
+  <div className="example-popover">
+    <Popover id="njdbt5">
+      <button className="btn">Open Popover</button>
+    </Popover>
+  </div>
+);
+
+export default ExamplePopover;