+++ src/examples/components/ExamplePopup/ExamplePopup.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import { Popup } from '../../../components';
+
+const ExamplePopup = () => (
+  <div className="example-popup">
+    <Popup id="njdbt5">
+      <button className="btn">Open Popup</button>
+    </Popup>
+  </div>
+);
+
+export default ExamplePopup;