+++ src/examples/components/ExamplePopup/index.js
@@ -1,0 +1,3 @@
+import ExamplePopup from './ExamplePopup';
+
+export default ExamplePopup;