+++ src/examples/components/ExampleSlider/ExampleSlider.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import { Slider } from '../../../components';
+
+const ExampleSlider = () => (
+  <div className="example-slider">
+    <Slider id="njdbt5">
+      <button className="btn">Open Slider</button>
+    </Slider>
+  </div>
+);
+
+export default ExampleSlider;