+++ src/examples/components/ExampleSlider/index.js
@@ -1,0 +1,3 @@
+import ExampleSlider from './ExampleSlider';
+
+export default ExampleSlider;