+++ src/examples/components/ExampleWidget/ExampleWidget.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import { Widget } from '../../../components';
+
+const ExampleWidget = () => (
+  <div className="example-widget">
+    <Widget id="njdbt5" height={400} />
+  </div>
+);
+
+export default ExampleWidget;