+++ src/examples/components/ExampleWidget/index.js
@@ -1,0 +1,3 @@
+import ExampleWidget from './ExampleWidget';
+
+export default ExampleWidget;