+++ src/examples/components/index.js
@@ -1,0 +1,7 @@
+import ExampleWidget from './ExampleWidget';
+import ExamplePopup from './ExamplePopup';
+import ExampleSlider from './ExampleSlider';
+import ExampleSidetab from './ExampleSidetab';
+import ExamplePopover from './ExamplePopover';
+
+export { ExampleWidget, ExamplePopup, ExampleSlider, ExampleSidetab, ExamplePopover };